backend/requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.12
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
flask-restful==0.3.10
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.35.2
setuptools==69.0.3
gunicorn==21.2.0
resend>=0.8.0
